If you are on a 64-bit platform and intptr_t is not the same as either int64_t or uint64_t it seems it may have difficulty deciding which one of those to widen to, and that can lead to a C++ moan about ambiguity. So here I put in a couple of explicit casts to force the issue.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@4998 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/arith12.cpp
@@ -281,7 +281,8 @@ LispObject large_modular_expt(LispObject a, int x)
 inline intptr_t muldivptr(uintptr_t a, uintptr_t b, uintptr_t c)
 {   if (!SIXTY_FOUR_BIT || c <= 0xffffffffU)
         return ((uint64_t)a*(uint64_t)b)%(uintptr_t)c;
-    else return (intptr_t)NARROW128((uint128(a)*uint128(a))%(uintptr_t)c);
+    else return (intptr_t)NARROW128((uint128((uint64_t)a)*
+                                     uint128((uint64_t)a))%(uintptr_t)c);
 }
 
 LispObject Lmodular_expt(LispObject env, LispObject a, LispObject b)
